1609: move storeConfig to general folder, get the categories depending on the buildConfig, fix test
