+++ administration/src/project-configs/koblenz/config.ts
@@ -2,6 +2,7 @@ import BirthdayExtension from '../../cards/extensions/BirthdayExtension'
 import RegionExtension from '../../cards/extensions/RegionExtension'
 import StartDayExtension from '../../cards/extensions/StartDayExtension'
 import { ProjectConfig } from '../getProjectConfig'
+import { storeConfig } from '../storeConfig'
 import { DataPrivacyBaseText, dataPrivacyBaseHeadline } from './dataPrivacyBase'
 import pdfConfig from './pdf'
 
@@ -26,9 +27,7 @@ const config: ProjectConfig = {
   cardStatistics: { enabled: false },
   freinetCSVImportEnabled: false,
   cardCreation: false,
-  storeManagement: {
-    enabled: false,
-  },
+  storeManagement: storeConfig,
   userImportApiEnabled: true,
 }
