+++ administration/src/project-configs/nuernberg/config.ts
@@ -4,11 +4,11 @@ import NuernbergPassIdExtension from '../../cards/extensions/NuernbergPassIdExte
 import RegionExtension from '../../cards/extensions/RegionExtension'
 import StartDayExtension from '../../cards/extensions/StartDayExtension'
 import { ProjectConfig } from '../getProjectConfig'
+import { storeConfig } from '../storeConfig'
 import ActivityLogEntry from './ActivityLogEntry'
 import { buildCsvLine } from './csvExport'
 import { DataPrivacyBaseText, dataPrivacyBaseHeadline } from './dataPrivacyBase'
 import pdfConfig from './pdf'
-import { storeConfig } from './storeConfig'
 
 const config: ProjectConfig = {
   name: 'Digitaler Nürnberg-Pass',
