+++ administration/src/util/getBuildConfig.ts
@@ -2,6 +2,8 @@ import {
   BAYERN_PRODUCTION_ID,
   BAYERN_STAGING_ID,
   BuildConfigType,
+  KOBLENZ_PRODUCTION_ID,
+  KOBLENZ_STAGING_ID,
   NUERNBERG_PRODUCTION_ID,
   NUERNBERG_STAGING_ID,
   buildConfigBayern,
@@ -17,6 +19,10 @@ export const getBuildConfig = (hostname: string): BuildConfigType => {
       return buildConfigBayern
     case NUERNBERG_PRODUCTION_ID:
     case NUERNBERG_STAGING_ID:
+      return buildConfigNuernberg
+    case KOBLENZ_PRODUCTION_ID:
+    case KOBLENZ_STAGING_ID:
+      // TODO 1647 Add correct buildConfig for koblenz
       return buildConfigNuernberg
     default:
       console.debug('Falling back to bayern.')
